REST, SpringDI & bean validation added. Tests not completed.

+++ src/main/java/pl/futurecollars/invoices/model/Vat.java
@@ -4,17 +4,17 @@ import java.math.BigDecimal;
 
 public enum Vat {
 
-    VAT_23(BigDecimal.valueOf(23)),
-    VAT_8(BigDecimal.valueOf(8)),
-    VAT_7(BigDecimal.valueOf(7)),
-    VAT_5(BigDecimal.valueOf(5)),
-    VAT_0(BigDecimal.valueOf(0)),
-    VAT_ZW(BigDecimal.valueOf(0));
+    VAT_23(23),
+    VAT_8(8),
+    VAT_7(7),
+    VAT_5(5),
+    VAT_0(0),
+    VAT_ZW(0);
 
     private BigDecimal rate;
 
-    Vat(BigDecimal rate) {
-        this.rate = rate;
+    Vat(int rate) {
+        this.rate = BigDecimal.valueOf(rate);
     }
 
     public BigDecimal getRate() {
